requirements.txt
setuptools
certifi
charset-normalizer
idna
numpy
oauthlib
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
six
tzdata
urllib3
atproto